+++ apps/block_scout_web/test/block_scout_web/views/tokens/holder_view_test.exs
@@ -40,6 +40,13 @@ defmodule BlockScoutWeb.Tokens.HolderViewTest do
 
       assert HolderView.total_supply_percentage(value, total_supply) == "0.0005%"
     end
+
+    test "zero total_supply" do
+      %Token{total_supply: total_supply} = build(:token, total_supply: 0)
+      %TokenBalance{value: value} = build(:token_balance, value: 0)
+
+      assert HolderView.total_supply_percentage(value, total_supply) == "N/A%"
+    end
   end
 
   describe "format_token_balance_value/1" do